# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beampipe",
    "description": "Beampipe Core\n\n## Getting started\n\n1. **Authenticate**: `POST /api/v1/login` with the admin **username or email** and password from setup. Copy `access_token` and click **Authorize** (Bearer token).\n2. **List projects**: `GET /api/v1/projects` shows registered project modules.\n3. **Register sources**: `POST /api/v1/sources` (returns 201 or 200 if already registered).\n4. **Run discovery**: `POST /api/v1/sources/discover` marks sources for async archive polling.\n5. **Create execution**: `POST /api/v1/executions/prepare` (preview) then `POST /api/v1/executions`.\n6. **Execute**: `POST /api/v1/executions/{id}/execute` enqueues staging/submit on the worker queue.\n7. **Poll**: `GET /api/v1/executions/{id}/status` or `.../ledger-snapshot` for operator view.\n\n### Common errors\n\n- **401**: missing or expired token. Log in again.\n- **400**: unknown `project_module` (check `GET /api/v1/projects`).\n- **404**: source, execution, or deployment profile not found.\n- **503**: Redis worker queue unavailable.\n\nHealth: `GET /api/v1/health` (liveness) \u00b7 `GET /api/v1/ready` (deps).\n\nLocal-only UI: [`/sources`](/sources) for a minimal browser view.",
    "version": "0.1.0",
    "contact": {
      "name": "admin",
      "email": "admin@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/deployment-profiles": {
      "get": {
        "tags": [
          "deployment-profiles"
        ],
        "summary": "List deployment profiles",
        "description": "List DALiuGE deployment profiles. Filter by project_module or global_only.",
        "operationId": "list_deployment_profiles_api_v1_deployment_profiles_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Items Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "project_module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project module; omit for all",
              "title": "Project Module"
            },
            "description": "Filter by project module; omit for all"
          },
          {
            "name": "global_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "When true, return only global (non-project) profiles",
              "default": false,
              "title": "Global Only"
            },
            "description": "When true, return only global (non-project) profiles"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse_DaliugeDeploymentProfileRead_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "deployment-profiles"
        ],
        "summary": "Create deployment profile",
        "description": "Create a DALiuGE deployment profile with translation and deployment configuration.",
        "operationId": "create_deployment_profile_api_v1_deployment_profiles_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DaliugeDeploymentProfileCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaliugeDeploymentProfileRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deployment-profiles/{profile_id}": {
      "get": {
        "tags": [
          "deployment-profiles"
        ],
        "summary": "Get deployment profile",
        "description": "Return a deployment profile by UUID.",
        "operationId": "get_deployment_profile_api_v1_deployment_profiles__profile_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaliugeDeploymentProfileRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "deployment-profiles"
        ],
        "summary": "Update deployment profile",
        "description": "Partially update a deployment profile. Nested translation/deployment fields are merged.",
        "operationId": "update_deployment_profile_api_v1_deployment_profiles__profile_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DaliugeDeploymentProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaliugeDeploymentProfileRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "deployment-profiles"
        ],
        "summary": "Delete deployment profile",
        "description": "Permanently delete a deployment profile.",
        "operationId": "delete_deployment_profile_api_v1_deployment_profiles__profile_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DaliugeDeploymentProfileCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Unique profile name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Human-readable description"
          },
          "project_module": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Module",
            "description": "Owning project module; null for global profiles"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "Use as default profile for the project or globally",
            "default": false
          },
          "translation": {
            "$ref": "#/components/schemas/DaliugeTranslationConfig"
          },
          "deployment": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RestRemoteDeploymentConfig"
              },
              {
                "$ref": "#/components/schemas/SlurmRemoteDeploymentConfig"
              }
            ],
            "title": "Deployment",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "rest_remote": "#/components/schemas/RestRemoteDeploymentConfig",
                "slurm_remote": "#/components/schemas/SlurmRemoteDeploymentConfig"
              }
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "translation",
          "deployment"
        ],
        "title": "DaliugeDeploymentProfileCreate",
        "description": "API create: nested ``translation`` + ``deployment`` only.",
        "examples": [
          {
            "deployment": {
              "kind": "rest_remote",
              "verify_ssl": false
            },
            "description": "Default REST remote profile for Setonix",
            "is_default": true,
            "name": "setonix-rest-default",
            "project_module": "wallaby_hires",
            "translation": {
              "algo": "metis",
              "num_islands": 0,
              "num_par": 1
            }
          }
        ]
      },
      "DaliugeDeploymentProfileRead": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Profile name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Human-readable description"
          },
          "project_module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Module",
            "description": "Owning project module when scoped"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "Default profile flag",
            "default": false
          },
          "translation": {
            "$ref": "#/components/schemas/DaliugeTranslationConfig",
            "description": "DALiuGE translator configuration"
          },
          "deployment": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RestRemoteDeploymentConfig"
              },
              {
                "$ref": "#/components/schemas/SlurmRemoteDeploymentConfig"
              }
            ],
            "title": "Deployment",
            "description": "REST or Slurm remote deployment configuration",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "rest_remote": "#/components/schemas/RestRemoteDeploymentConfig",
                "slurm_remote": "#/components/schemas/SlurmRemoteDeploymentConfig"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "name",
          "translation",
          "deployment"
        ],
        "title": "DaliugeDeploymentProfileRead",
        "description": "API read."
      },
      "DaliugeDeploymentProfileUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "project_module": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Module"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "translation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DaliugeTranslationPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "deployment": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment",
            "description": "Deployment config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DaliugeDeploymentProfileUpdate"
      },
      "DaliugeTranslationConfig": {
        "properties": {
          "algo": {
            "type": "string",
            "enum": [
              "metis",
              "mysarkar"
            ],
            "title": "Algo",
            "description": "Partition algorithm",
            "default": "metis"
          },
          "num_par": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Num Par",
            "description": "Partitions / nodes",
            "default": 1
          },
          "num_islands": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Num Islands",
            "description": "Data islands; dlg -s",
            "default": 0
          },
          "tm_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Tm Url",
            "description": "Translator base URL"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DaliugeTranslationConfig",
        "description": "DALiuGE translator (partition + TM URL)."
      },
      "DaliugeTranslationPatch": {
        "properties": {
          "algo": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "metis",
                  "mysarkar"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Algo"
          },
          "num_par": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Par"
          },
          "num_islands": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Islands"
          },
          "tm_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Tm Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DaliugeTranslationPatch",
        "description": "Partial translation."
      },
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable error message",
            "examples": [
              "Resource not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Standard FastAPI / HTTPException error body."
      },
...
      "PaginatedListResponse_DaliugeDeploymentProfileRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DaliugeDeploymentProfileRead"
            },
            "type": "array",
            "title": "Data"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "items_per_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items Per Page"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total_count",
          "has_more"
        ],
        "title": "PaginatedListResponse[DaliugeDeploymentProfileRead]"
      },
...
      "RestRemoteDeploymentConfig": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "rest_remote",
            "title": "Kind",
            "default": "rest_remote"
          },
          "dim_host_for_tm": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Dim Host For Tm",
            "description": "DIM host as seen by translator (gen_pg)"
          },
          "dim_port_for_tm": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dim Port For Tm"
          },
          "deploy_host": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Deploy Host"
          },
          "deploy_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Deploy Port"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RestRemoteDeploymentConfig",
        "description": "DALiuGE REST remote deployment"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login"
          }
        }
      }
    }
  },
...
  "servers": [
    {
      "url": "/",
...
    }
  ],
...
}